fix: int is equivalent in dart

+++ lib/domain/entity/user.dart
@@ -2,7 +2,7 @@
 import 'dart:ffi';
 
 class User {
-  final Long id;
+  final int id;
   final String name;
   final Gender gender;
   final String email;
